Make wav-copy accept both xspecifiers and xfilenames
In scripts such as perturb-speed and perturb-volume scp lines are
tranformed into piped command with the appropropriate sox command. The
case that the scp file has file offsets was not handled. This commit
both generalizes the wav-copy command to work also on xfilenames and
fixes the two perturb scripts to use this command in case of file
offsets.