[src] Fix two bugs in batched-wav-nnet3-cuda binary. (#3600)
1)  Using the "Any" apis prior to finishing submission of the full group
could lead to a group finishing early.  This would cause output to
appear repeatedly.  I have not seen this occur but an audit revealed it
as an issue. The fix is to use the "Any" APIs only when full groups have
been submitted.

2) GetNumberOfTasksPending() can return zero even though groups have not
been waited for.  This API call should not be used to determine if all
groups have been completed as the number of pending tasks is independent
of the number of groups remaining.